feature alias 추가, navigation.jsx 요소별 컴포넌트 화
