+++ src/App.css
@@ -34,11 +34,20 @@ body {
 }
 
 body.day {
-  color: #212121 ;
+  color: #212121;
   background-color: white;
 }
 
 /* 스크롤바 안보이게 ( 크롬, 사파리, 오페라, 엣지 ) 동작 */
 body::-webkit-scrollbar {
   display: none;
+}
+
+/* LINK태그 기본 밑줄 등 꾸밈요소 삭제 */
+a {
+  text-align: center;
+  text-decoration: none;
+  /* 링크의 밑줄 제거 */
+  color: inherit;
+  /* 링크의 색상 제거 */
 }
