+++ src/style/Login.css
@@ -1,3 +1,4 @@
+@import url(//fonts.googleapis.com/earlyaccess/notosanskr.css);
 @font-face {
   font-family: 'GmarketSansMedium';
   src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2001@1.1/GmarketSansMedium.woff')
@@ -6,7 +7,6 @@
   font-style: normal;
 }
 
-@import url(//fonts.googleapis.com/earlyaccess/notosanskr.css);
 
 /* 모든 요소에 대해 box-sizing 속성을 border-box로 설정 */
 * {
