+++ src/style/SideBar.css
@@ -106,7 +106,7 @@
   text-align: center;
 
   position: absolute;
-  width: 180px;
+  width: 15rem;
   margin: -100px 0 0 -50px;
   padding: 120px 50px 50px 63px;
   border-radius: 20px;
@@ -124,7 +124,7 @@
 #leftside li
 {  text-align: left;
   padding: 10px 0;
-  font-size: 16px;
+  font-size: 1rem;
 }
 
 /*
